requirements.txt
ultralytics
supervision
numpy
pillow
opencv-python
requests

torch>=2.3.0+cu121
torchaudio>=2.3.0+cu121
torchvision>=0.18.0
--index-url https://download.pytorch.org/whl/cu121
